sort incoming shares before doing anything else
there are a bunch of insidious assumptions we use in matrix construction
and multiplication that cause the syndrome fast past calculations to
be wrong.

let's just always sort them first: decode already does this at the top
too, which is why we avoided it here, but it makes everything less likely
to blow up at low/reasonable cost.