Avoid errors.ProgrammerError class (#7)
The errors.ProgrammerError class is defined in spacemonkeygo/errors with
the LogOnCreation option. This forces the error to be logged with the
stack trace at the moment it is created. This might be undesired by apps
using the infectious library as it polutes the log output. For example,
this happens to me in the build as I have a test suite covering
erroneous use cases.

This patch replaces the errors.ProgrammerError class with the already
defined Error class in the infectious library.